v1

+++ data/build_hbs.py
@@ -1,16 +1,18 @@
+#! /home/extradisk/linchenran/.pyenv/versions/matlab/bin/python
 import os
-
+import click
+from loguru import logger
 import matlab.engine
+from genericpath import exists
 
 from utils.timeout import timeout
 
+TIMEOUT = 60
 
 def init_setting(eng):
     matlab_dir = '/home/extradisk/linchenran/hbs_seg'
     init_script = f'''
-    clear all;
-    close all;
-    cd('{matlab_dir}');
+    addpath('{matlab_dir}');
     addpath('{matlab_dir}/dependencies');
     addpath('{matlab_dir}/dependencies/im2mesh');
     addpath('{matlab_dir}/dependencies/mfile');
@@ -18,40 +20,56 @@ def init_setting(eng):
     '''
     eng.eval(init_script, nargout=0)
 
-@timeout(60)
-def generate_hbs(eng, image_path, timeout=30):
+@timeout(TIMEOUT)
+def generate_hbs(eng, image_path, mat_path):
     hbs_script = f'''
     hbs = compute_hbs_from_image('{image_path}');
-    hbs_dict.{image_path.split('/')[-1].replace('.', '_')} = hbs;
+    save('{mat_path}', 'hbs');
     '''
-    print(f'Load image {image_path} and try to generate HBS...')
+    logger.info(f'Load image {image_path}')
     eng.eval(hbs_script, nargout=0)
-    print('Generated HBS successfully')
+    logger.info(f'Generated HBS to {mat_path}')
 
 
-def save_hbs_dict(eng, mat_path):
-    save_script = f'''
-    save('{mat_path}', 'hbs_dict');
-    '''
-    eng.eval(save_script, nargout=0)
-
-def main():
-    image_dir = '/home/extradisk/linchenran/hbs_torch/img/generated'
-    eng_shared_name = 'hbs_engine'
-    eng = matlab.engine.start_matlab(f'-nojvm -nodisplay -nosplash -nodesktop -r "matlab.engine.shareEngine(\'{eng_shared_name}\')"')
-
+def start_eng(shared_name='hbs_engine'):
+    eng = matlab.engine.start_matlab(f'-nojvm -nodisplay -nosplash -nodesktop -r "matlab.engine.shareEngine(\'{shared_name}\')"')
     init_setting(eng)
-    for file in os.listdir(image_dir):
-        if file.endswith('.png'):
-            image_path = os.path.join(image_dir, file)
-            try:
-                generate_hbs(eng, image_path)
-            except TimeoutError:
-                os.remove(image_path)
-                print(f"Failed and remove image {image_path}")
+    logger.info('Started matlab eng')
+    return eng
 
-    save_hbs_dict(eng, f'{image_dir}/hbs_dict.mat')
+def restart_eng(eng, shared_name='hbs_engine'):
+    logger.warning('Restarting matlab eng')
     eng.quit()
-    
+    eng = start_eng(shared_name)
+    return eng
+
+
+@click.command()
+@click.option("--image_dir", default='/home/extradisk/linchenran/hbs_torch/img/generated', help="Directory of images to generate HBS")
+@click.option("--eng_name", default='hbs_engine', help="Shared name of the matlab engine")
+def main(image_dir, eng_name):
+    eng = start_eng(eng_name)
+
+    for file in os.listdir(image_dir):
+        try:
+            if file.endswith('.png'):
+                image_path = os.path.join(image_dir, file)
+                mat_path = image_path.replace('.png', '.mat')
+                if exists(mat_path):
+                    logger.info(f'HBS {mat_path} already exists')
+                    continue
+
+                try:
+                    generate_hbs(eng, image_path, mat_path)
+                except TimeoutError:
+                    eng = restart_eng(eng, eng_name)
+                    os.remove(image_path)
+                    logger.error(f"Timeout, failed to generate and removed image {image_path}")
+        except Exception as e:
+            logger.error(e)
+            continue
+
+    eng.quit()
+
 if __name__ == "__main__":
     main()
